+++ src/components/common/ListWithIcons/ListWithIcons.types.ts
@@ -1,0 +1,9 @@
+import { ComponentType } from 'react'
+
+export interface ListWithIconsItemData {
+  icon?: ComponentType<{ fillColor?: string; size?: number }>
+  text?: string
+  disabled?: boolean
+  tooltip?: string
+  textColor?: string
+}
